requirements.txt
pandas>=1.5
scikit-learn>=1.0
Pillow>=9.0
